Refactor fscore outside goban.rs to fscore.rs
